move command classes to enable usage of config path
