+++ contribs/application/src/main/java/org/matsim/application/RunScenario.java
@@ -1,12 +1,11 @@
-package org.matsim.application.commands;
+package org.matsim.application;
 
-import org.matsim.application.MATSimApplication;
 import picocli.CommandLine;
 
 import java.util.concurrent.Callable;
 
 @CommandLine.Command(name = "run", description = "Run the scenario")
-public class RunScenario implements Callable<Integer> {
+class RunScenario implements Callable<Integer> {
 
     @CommandLine.ParentCommand
     private MATSimApplication app;
